requirements.txt
discord.py==1.7.3
requests==2.26.0
beautifulsoup4==4.10.0
discord==1.7.3
Pillow==8.4.0
